# api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "RecipeDB 3.0 API",
    "description": "API Documentation for RecipeDB 3.0 Backend.\n\n## Important Rules\n1. **ID Padding**: All IDs (`Recipe_ID`, `Ing_ID`) are **8-digit zero-padded strings** (Except FlavorDB_ID). \n    * Example: Input `1` is automatically treated as `00000001`.\n2. **Pagination**: Most list endpoints accept a `page` parameter (default: 1). The page size is fixed at **7 items**.\n3. **Images**: Images are returned as URLs. If a recipe has no image, the field may be empty or missing.",
    "version": "2.1.0",
    "contact": {
      "name": "RecipeDB Team",
      "email": "bagler@iiitd.ac.in"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3033/api/recipes",
      "description": "Local Development Server"
    }
  ],
  "paths": {
...
    "/by-ingredients": {
      "get": {
        "summary": "Filter by Ingredients",
        "description": "**Source File:** `by-ingredients.js`\n\n### What it does\nFinds recipes containing specific ingredients.\n\n### When to use\nUse for **Fridge Search** to display all the recipes containing and excluding particular ingredients.\n\n### Parameters\n* **include** (Optional): string (\"onion,garlic\", \"tomato,potato\")\n* **exclude** (Optional): string (\"beef\", \"eggs\", \"milk\")",
        "parameters": [
          { "name": "include", "in": "query", "schema": { "type": "string", "example": "onion,garlic" } },
          { "name": "exclude", "in": "query", "schema": { "type": "string", "example": "beef" } }
        ],
        "responses": { "200": { "description": "List of recipes." } }
      }
    },
    "/by-ingredients-categories-title": {
      "get": {
        "summary": "Mega Filter (Advanced Search)",
        "description": "**Source File:** `by-ingredients-categories-title.js`\n\n### What it does\nCombines Ingredient, Diet, Flavor, Category, Cuisine, and Title filters.\n\n### When to use\nUse for **Advanced Search**.\n\n### Parameters\n* **includeDiet** (Optional): string (\"Vegetarian\", \"Vegan\")\n* **title** (Optional): string (\"Salad\", \"Curry\")\n* **includeFlavor** (Optional): string (\"Spice\", \"Herb\", \"Fruit\")\n* **excludeFlavor** (Optional): string (\"Bitter\", \"Sour\")\n* **category** (Optional): string (\"Lunch\", \"Breakfast\")\n* **includeIngredient** (Optional): string (\"garlic\", \"onion\")\n* **excludeIngredient** (Optional): string (\"beef\")\n* **cuisine** (Optional): string (\"American\")",
        "parameters": [
          { "name": "includeDiet", "in": "query", "schema": { "type": "string", "example": "Vegetarian" } },
          { "name": "title", "in": "query", "schema": { "type": "string", "example": "Salad" } },
          { "name": "includeFlavor", "in": "query", "schema": { "type": "string", "example": "Spice" } },
          { "name": "excludeFlavor", "in": "query", "schema": { "type": "string", "example": "Bitter" } },
          { "name": "category", "in": "query", "schema": { "type": "string", "example": "Lunch" } },
          { "name": "includeIngredient", "in": "query", "schema": { "type": "string", "example": "garlic" } },
          { "name": "excludeIngredient", "in": "query", "schema": { "type": "string", "example": "beef" } },
          { "name": "cuisine", "in": "query", "schema": { "type": "string", "example": "American" } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1, "example": 1 } }
        ],
        "responses": { "200": { "description": "List of recipes." } }
      }
    },
...
    "/atlas-search/autocomplete": {
      "get": {
        "summary": "Search Autocomplete (Atlas Search)",
        "description": "**Source File:** `atlas-search.js`\n\n### What it does\nProvides lightning-fast, typo-tolerant autocomplete suggestions for recipe titles using MongoDB Atlas Search (Lucene engine).\n\n### When to use\nIntegrate this directly into the Navbar/Header Search Bar to show a dropdown of 10 quick suggestions as the user types.",
        "parameters": [
          { "name": "q", "in": "query", "required": true, "schema": { "type": "string", "example": "chic" } }
        ],
        "responses": { "200": { "description": "Array of top 10 lightweight recipe objects (ID, Title, Image)." } }
      }
    },
    "/atlas-search/search": {
      "get": {
        "summary": "Full Fuzzy Search (Atlas Search)",
        "description": "**Source File:** `atlas-search.js`\n\n### What it does\nPerforms a full text search on recipe titles with advanced fuzzy matching (typo tolerance). E.g., searching for 'chikcen' will successfully return 'chicken' recipes.\n\n### When to use\nUse this to power the main Search Results Page when a user hits 'Enter' on their query.",
        "parameters": [
          { "name": "q", "in": "query", "required": true, "schema": { "type": "string", "example": "chikcen" } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1, "example": 1 } }
        ],
        "responses": { "200": { "description": "Paginated array of detailed recipe objects." } }
      }
    }
  }
}